% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMandala.R
\name{PlotMandala}
\alias{PlotMandala}
\title{Plot Mandala}
\usage{
PlotMandala(
  radius = c(0.6, 1.8),
  npoints = 14L,
  depth = 3L,
  scheme = NULL,
  seed = NULL
)
}
\arguments{
\item{radius}{'numeric' vector of length 1 or 2.
Factor of expansion or compression.
A random-number generator may be used by defining the
lower and upper limits of a uniform distribution.}

\item{npoints}{'integer' count.
Number of points}

\item{depth}{'integer' count.
Number of iterations}

\item{scheme}{'character' vector.
Name of color palette(s) to choose from, see
\code{\link[wesanderson:wes_palette]{wes_palette}} function for choices.
By default, specified using a random selection from all possible palettes.}

\item{seed}{'integer' count.
Random number generator state, used to replicate the results.}
}
\value{
Invisible \code{NULL}
}
\description{
Draw a colored mandala using Voronoi tessellation and a Wes Anderson color palette.
}
\examples{
PlotMandala(seed = 222)

\dontrun{
grDevices::png("mandala.png")
PlotMandala()
grDevices::dev.off()
}

}
\references{
This function was derived from the
  \href{https://github.com/aschinchon/mandalas-colored}{mandalas-colored}
  script by A.S. Chinchón, accessed on Dec 13, 2019.
}
\author{
J.C. Fisher
}
\keyword{hplot}
